move SimulationManager into invertedai/ file out of helpers/
