+++ invertedai/__init__.py
@@ -54,7 +54,7 @@ from invertedai.logs.debug_logger import DebugLogger
 
 # --- Helpers ---
 from invertedai.helpers.waypoints import WaypointManagerConfig, WaypointManager
-from invertedai.helpers.simulation_manager import SimulationManager
+from invertedai.simulation_manager import SimulationManager
 
 # --- Common data types ---
 from invertedai.common import (
