+++ invertedai/helpers/__init__.py
@@ -6,5 +6,4 @@ __all__ = [
     "WaypointManagerConfig",
     "SceneVisualizer",
     "SceneVisualizerConfig",
-    "SimulationAgentDict,"
 ]
