+++ invertedai/simulation_manager.py
